@@ -1,5 +1,5 @@
 import {Request, Response} from "express";
-import paypal = require('paypal-rest-sdk')
+import paypal from 'paypal-rest-sdk';
 
 paypal.configure({
     mode: 'sandbox',
