@@ -1,4 +1,4 @@
-const nodemailer = require('nodemailer');
+import nodemailer from 'nodemailer'
 
 export const sendEmail = async ({ to, subject, text }) => {
     try {
