@@ -1,8 +1,8 @@
 import { Request, Response } from 'express'
-import gDB from '../InitDataSource'
-import { UserEntity } from '../entity/User.entity'
 import { validate } from 'class-validator'
-import { CLog } from '../AppHelper'
+import { CLog } from '../AppHelper.js'
+import gDB from '../InitDataSource.js'
+import { UserEntity } from '../entity/User.entity.js'
 
 class UserController {
   static db = gDB.getRepository(UserEntity)
