@@ -1,5 +1,5 @@
 import {Router} from "express";
-import {CartController} from "../controllers/CartController";
+import {CartController} from "../controllers/CartController.js";
 
 const router = Router()
 
